Implementa toda funcionalidade de explorador de arquivos no workspace.

Este PR corrige um milhão de bugs, reportados a partir da
[issue](https://github.com/COCSI-MG/eevee/issues/105).

Simplifica a lib filestash que é responsável pelo uso do indexedb no
navegador.

+++ CEFETCodeLab-SchedulerApi/src/user/user.service.ts
@@ -48,7 +48,16 @@ export class UserService {
     return this.userRepository.findOne({ where: { email } });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const user = await this.findOne(id);
+    if (!user) {
+      throw new Error('User not found');
+    }
+
+    if (user.isAdmin) {
+      throw new Error('Cannot delete admin user');
+    }
+
     return this.userRepository.delete({ id });
   }
 }
